# Fixes #34160 - Improve fact_values api performance