--- a/app/models/fact_value.rb
+++ b/app/models/fact_value.rb
@@ -88,17 +88,6 @@ class FactValue < ApplicationRecord
     output
   end
 
-  def self.build_facts_hash(facts)
-    hosts = Host.where(:id => facts.group_by(&:host_id).keys).all
-
-    hash = {}
-    facts.each do |fact|
-      hash[hosts.detect { |h| h.id == fact.host_id }.to_s] ||= {}
-      hash[hosts.detect { |h| h.id == fact.host_id }.to_s].update({fact.name.to_s => fact.value})
-    end
-    hash
-  end
-
   # converts all strings with units (such as 1 MB) to GB scale and Sum them
   # returns an array with total sum and number of elements
   def self.to_gb(fact)