Previously, the fact_values api called loaded a lot of un-needed data
from the database and used a sub-optimal way of converting the data to a
hash for display.
This commit changes the implementations so that we only load the
required columns from the database and generate the hash in a much more
efficient way. It also moves the build_facts_hash method from FactValue
model to the controller, which is its only consumer.

Some benchmarks:
Before, loading 20 facts (default per_page setting):
Completed 200 OK in 52ms (Views: 2.6ms | ActiveRecord: 15.5ms | Allocations: 35603)
Before, loading 1000 facts:
Completed 200 OK in 1181ms (Views: 2.2ms | ActiveRecord: 22.8ms | Allocations: 1390471)
Before, loading 10000 facts:
Completed 200 OK in 14581ms (Views: 7.4ms | ActiveRecord: 75.6ms | Allocations: 22283568)

After, loading 20 facts:
Completed 200 OK in 37ms (Views: 3.8ms | ActiveRecord: 18.3ms | Allocations: 15549)
After, loading 1000 facts:
Completed 200 OK in 53ms (Views: 2.4ms | ActiveRecord: 16.0ms | Allocations: 47083)
After, loading 10000 facts:
Completed 200 OK in 313ms (Views: 5.8ms | ActiveRecord: 38.7ms | Allocations: 433106)


<!---

Thank you for contributing to The Foreman, please read the
[following guide](https://www.theforeman.org/contribute.html), in short:

* [Create an issue](https://projects.theforeman.org/projects/foreman/issues)
* Reference the issue via `Fixes #1234` in the commit message
* Prefer present-tense, imperative-style commit messages
* Mark all strings for translation, see [1]
* Suggest prerequisites for testing and testing scenarios following example above.
* Prepend `[WIP]` for work in progress to prevent bots from triggering actions
* Be patient, we will do our best to take a look as soon as we can
* Explain the purpose of the PR, attach screenshots if applicable
* List all prerequisites for testing (e.g. VMware cluster, two smart proxies...)
* Reviewers often use extensive list of items to check, have a look prior submitting [2]
* Be nice and respectful

1: https://projects.theforeman.org/projects/foreman/wiki/Translating#Translating-for-developers
2: https://github.com/theforeman/foreman/blob/develop/developer_docs/pr_review.asciidoc
-->